+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@
 - Added help icons for the fields "Game Name" and "Game File Name" in the table data dialog with hints about the renaming that results out of changing this fields.
 - Added all possible uploads into one split-button menu. This way, the you don't have to search for the uploads buttons in the different sections anymore.
 
-<img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/tables/upload-button.png" width="400" />
+<img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/tables/upload-buttons.png" width="300" />
 
 
 ### Bugfixes
@@ -15,4 +15,5 @@
 - Fixed possible database lock when Popper is running and Studio is started with new tables detected.
 - Added missing .mp4 and .apng file formats to the list of possible drag and drop files for the assets dialog.
 - Improved emulator start/exit script analysis when writing the required "curl" calls for the VPin Studio server.
+- Fixed Table Upload dialog: "Upload and Replace" and "Upload and Clone" options are disabled when opened without a table selection.
 - Added missing automatic ".ini" file renaming when the "Game File Name" is renamed.